requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
pillow==11.2.0
PySide6==6.9.0
PySide6_Addons==6.9.0
PySide6_Essentials==6.9.0
qrcode==8.1
requests==2.32.3
shiboken6==6.9.0
urllib3==2.3.0
